Use a single script for CI
We often end up wanting to split a test suite so it can be run in
parallel but how we split depends somewhat on the CI server and the
kinds of tests in the suite e.g. it may make sense to split by test type
or by tag. Running all the tests in one command is the most
straightforward way to begin with so we are changing the default to
encourage that.

This change also avoids having to do complicated things to calculate
test coverage.