Create 'backend-base' variant and move files from project root into it
This refactors the code to make it easier to tell which files will be
copied into the resulting Rails app and which files are for the development
of this template.

We had to make copies of some files (e.g. .eslintrc.js) which were being
used by this template repo for development and also copied into the Rails
app.