Fix intermittent failure in user sign in spec (#257)
* Fix intermittent failure in user sign in spec

Devise assigns rememberable data to a signed cookie. Signed cookies are generated using ActiveSupport::MessageVerifier.generate, which encodes the string as the base64 encoded signed data and the signature for the data, separated by double dashes ('--'). Parsing the entire message as base64 can intermittently fail as the signature is not valid base64 encoded, resulting in extra bytes being added to the decoded string, which causes JSON.parse to fail.

* Remove trailing whitespace

Co-authored-by: Eoin Kelly <eoin@eoinkelly.info>