Configure I18n when translations are missing in test mode
Quite a few of our applications use I18n to try and separate content from code. Quite often, we end up accidentally using a fallback translation value, when we intended for an actual value to be present. To defend against that, often our tests will use `I18n.t!` to raise when it can't find a value. I think a more elegant solution would be just to tell ActionView to raise an error at that point when our translations are missing, rather than rely on the test writer to check.

I would be open to adding this to development.rb as well, but I'm unsure how much it would interfere with trying to prototype up a template for a feature.