Load the template config YAML file from the parent directory (#406)
The README describes placing the config YAML file in your working
directory, and running `rails new ....` to allow the template to pick up
the config file.

The problem with this, as far as I can tell, is that `template.rb` is
evaluated in the context of the directory `rails new` has just created,
which does not contain the config file. I think the solution is to
traverse up a level and load the config file from there. This won't work
if the user has provided an absolute path, but in that case
`ENV["CONFIG_PATH"]` can be provided to override it.

As an alternative workaround, a user could create the directory for
their app, place the config file inside it, and then run `rails new`
(without this patch)

---------

Co-authored-by: Gareth Jones <gareth.jones@ackama.com>
Co-authored-by: Eoin Kelly <eoin@eoinkelly.info>
Co-authored-by: Luke <lukedavia@icloud.com>
Co-authored-by: Eoin Kelly <eoin.kelly@ackama.com>