feat: support Rails 7.1 (#502)
Resolves #500 

This aims to be the minimal changes required for the template to work
with Rails 7.1

### Things which are not in scope for this PR

* Rails now generates docker related files - I've forced/skipped those
because their docker is focused on production and the diff is quite
large so I think it's best to tackle that as its own separate thing

### TODO
- [x] Do the basic upgrade
- [x] Adjust `frontend-base` to account for Shakapacker now doing a
thing we previously did ourselves.
- [x] `Rack::Handler is deprecated and replaced by Rackup::Handler`
warning. This appears during system tests. It is a knownw [Capybara
issue which is fixed but waiting on a gem
release](https://github.com/teamcapybara/capybara/pull/2706). Since this
is just a warning and will go away in due course, we can ignore it.
- [x] Resolve `Setting 'useContentHash' to 'false' in the production
environment (specified by NODE_ENV environment variable) is not allowed!
Content hashes get added to the filenames regardless of setting
useContentHash in 'shakapacker.yml' to false.` warning. This warning
seems to caused by the default `shakapacker.yml`. Shakapacker generates
a `shakapacker.yml` file where `useContentHash: false` is set in the
`&default` section and then `useContentHash: true` is explicitly set in
the production section. It doesn't seem like a real problem so I think
we can ignore this.
- [x] Make a decision about #507 and implement it here 
    - We have decided to leave this option commented out for now.
- [ ] Make a decision about how to handle the
`Rails.application.secrets` deprecation because it will be removed in
7.2. https://github.com/ackama/rails-template/issues/341 has good
context on the last time we discussed this.

---------

Co-authored-by: Eoin Kelly <eoin@eoinkelly.info>
Co-authored-by: Eoin Kelly <eoin.kelly@ackama.com>