Centralise and improve Rails version handling (#561)
Create a central place to manage which versions of Ruby and Rails this
template supports.

This change introduces a new configuration file called
`target_verions.yml`. This file holds data about which versions of Ruby
and Rails are currently compatible with the template.

The data is used in multiple contexts:

1. The template itself uses it to verify that the version of Rails being
used to run the template is compatible
2. CI uses it to choose a compatible version of Rails to install
3. Rubocop configuration for this repo uses it to ensure that Rubocop
checks which run against this repo will be compatible with all versions
of Ruby and Rails which the template supports. The Rubocop configuration
in the generated apps does not need this information because it can
infer it from the specific versions of Ruby and Rails in use in that
app.

This change also 

* updates CI to always choose the most recent Rails version compatible
with this template
* Gets CI passing - closes issue #550