feat: require all partials use strict locals (#594)
[Strict
locals](https://guides.rubyonrails.org/action_view_overview.html#strict-locals)
are a special comment that explicitly declares what local variables an
ERB partial accepts, along with default values for when an optional
variable is not present.

This improvements performance when dealing with large collections and
applications by making partials monomorphic (rather than polymorphic,
which can potentially turn megamorphic) resulting in them being compiled
only once regardless of the variations in their parameters; otherwise, a
new version of the partial has to be compiled for each different set of
parameters they're rendered with.

Additionally, these comments make it easier to understand the partial
and for both humans and tools to catch silly mistakes, since the comment
defines exactly what locals do and do not exist within the partial.

I expect in general this shouldn't be too painful, though we will need
to handle this for partials that come from third-parties such as
ActionText and Devise - my experience this so far has been very
painless, and any mistakes should be caught by our test suites quite
easily (as generally it just requires the partial to be rendered, even
if nothing actually asserts the content).

Finally, I believe this form of the comment can serve as the ultimate
bail out, allowing any local to be passed while still technically using
strict locals:

```
<%# locals: (**attributes) %>
```